Revert "added other time conversion functions to the file formerly known as humanToSeconds"
This reverts commit 05f078dbb63b334fd13b4471117e7dc80c451f8d.
